requirements.txt

flask
langchain
pinecone-client
sentence-transformers
ctransformers
pypdf
python-dotenv
-e .
